Finish with menu 1-8 and 16, 19

+++ Project/Product.h
@@ -50,8 +50,8 @@ public:
 
 	virtual bool operator ==(const Product& other) const = 0;
 
-	float totalCostOfGoods() const;
-	float totalRevenue() const;
+	double totalCostOfGoods() const;
+	double totalRevenue() const;
 	int endOfDayInventory() const;
 
 	virtual string showInfo() const = 0;
